requirements.txt
# python -m venv ./venv
# source ./venv/bin/activate
# pip install -r ./requirements.txt
# python3 -m ipykernel install --name=venv
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
gspread
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
numpy==1.24.0
oauthlib==3.2.2
pandas
pandas-gbq
pendulum
plotly==5.6.0
pycausalimpact==0.1.1
pygsheets
urllib3==1.26.6